fix: tweak conflicts select to handle nulls in the where conditions
You can't just invert (x=y) to NOT(x=y) if x might be null. It has to be (NOT(x=y) OR X is NULL).
Turns out to be easier to generate NOT(x=y and x is not null). It reads a little bit oddly, but the
logic is exactly equivalent.